create push file
